backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-pptx==1.0.2
python-docx==1.1.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
aiofiles==24.1.0
